move ReferencesOperation family classes under common directory
