+++ php/v2/src/ReferencesOperation/Factory/ReferencesOperationFactory.php
@@ -3,10 +3,10 @@
 namespace NodaSoft\ReferencesOperation\Factory;
 
 use NodaSoft\DataMapper\Factory\MapperFactory;
-use NodaSoft\OperationParams\ReferencesOperationParams;
+use NodaSoft\ReferencesOperation\Params\ReferencesOperationParams;
 use NodaSoft\ReferencesOperation\Command\ReferencesOperationCommand;
 use NodaSoft\Request\Request;
-use NodaSoft\Result\Operation\ReferencesOperation\ReferencesOperationResult;
+use NodaSoft\ReferencesOperation\Result\ReferencesOperationResult;
 
 interface ReferencesOperationFactory
 {
