+++ php/v2/src/ReferencesOperation/Params/TsReturnOperationParams.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace NodaSoft\OperationParams;
+namespace NodaSoft\ReferencesOperation\Params;
 
+use NodaSoft\ReferencesOperation\Params\ReferencesOperationParams;
 use NodaSoft\Request\Request;
 
 class TsReturnOperationParams implements ReferencesOperationParams
